# Reduce allocations when calling BeginAbsoluteSequence()