Used very liberally in osu! hitobjects.

`ThreadStatic` is used because this can be used from:
- Either of the BDL load threads.
- The update thread.

The cost of `ThreadStatic` becomes basically negligible in .NET 8, but this pathway is not hot enough to cause performance issues on its own: https://github.com/smoogipoo/Benchmarks/blob/37dc7c07b16cc1dd4898987a06c43cce438150c9/BenchmarksProject/BenchmarkThreadStatic.cs

Before:

|                     Method |     Mean |   Error |  StdDev |    Gen0 | Allocated |
|--------------------------- |---------:|--------:|--------:|--------:|----------:|
|               NonRecursive | 972.8 us | 1.89 us | 1.68 us |       - |     187 B |
|                  Recursive | 972.7 us | 2.70 us | 2.52 us |  9.7656 |   24259 B |
| SlightlyNestedNonRecursive | 972.2 us | 2.94 us | 2.75 us |       - |     187 B |
|    SlightlyNestedRecursive | 976.0 us | 2.28 us | 2.13 us | 33.2031 |   72331 B |
|     VeryNestedNonRecursive | 975.1 us | 5.91 us | 5.24 us |       - |     187 B |
|        VeryNestedRecursive | 975.9 us | 1.84 us | 1.72 us | 23.4375 |   49555 B |

After:

|                     Method |     Mean |   Error |  StdDev | Allocated |
|--------------------------- |---------:|--------:|--------:|----------:|
|               NonRecursive | 976.6 us | 3.80 us | 3.56 us |     187 B |
|                  Recursive | 986.9 us | 1.19 us | 1.06 us |     178 B |
| SlightlyNestedNonRecursive | 971.7 us | 2.77 us | 2.59 us |     187 B |
|    SlightlyNestedRecursive | 974.8 us | 3.51 us | 3.11 us |     179 B |
|     VeryNestedNonRecursive | 973.6 us | 2.21 us | 1.96 us |     187 B |
|        VeryNestedRecursive | 973.4 us | 1.76 us | 1.64 us |     179 B |